+++ src/main/java/com/raoulvdberge/refinedstorage/RSItems.java
@@ -24,4 +24,5 @@ public final class RSItems {
     public static final ItemCover COVER = new ItemCover();
     public static final ItemHollowCover HOLLOW_COVER = new ItemHollowCover();
     public static final ItemWrench WRENCH = new ItemWrench();
+    public static final ItemProcessorBinding PROCESSOR_BINDING = new ItemProcessorBinding();
 }