+++ src/main/java/com/raoulvdberge/refinedstorage/item/ItemProcessor.java
@@ -18,7 +18,6 @@ public class ItemProcessor extends ItemBase {
     public static final int TYPE_BASIC = 3;
     public static final int TYPE_IMPROVED = 4;
     public static final int TYPE_ADVANCED = 5;
-    public static final int TYPE_CUT_SILICON = 6;
 
     public ItemProcessor() {
         super(new ItemInfo(RS.ID, "processor"));
@@ -32,22 +31,20 @@ public class ItemProcessor extends ItemBase {
     public void registerModels(IModelRegistration modelRegistration) {
         modelRegistration.setModelVariants(
             this,
-            new ResourceLocation(RS.ID, "cut_basic_processor"),
-            new ResourceLocation(RS.ID, "cut_improved_processor"),
-            new ResourceLocation(RS.ID, "cut_advanced_processor"),
+            new ResourceLocation(RS.ID, "raw_basic_processor"),
+            new ResourceLocation(RS.ID, "raw_improved_processor"),
+            new ResourceLocation(RS.ID, "raw_advanced_processor"),
             new ResourceLocation(RS.ID, "basic_processor"),
             new ResourceLocation(RS.ID, "improved_processor"),
-            new ResourceLocation(RS.ID, "advanced_processor"),
-            new ResourceLocation(RS.ID, "cut_silicon")
+            new ResourceLocation(RS.ID, "advanced_processor")
         );
 
-        modelRegistration.setModel(this, TYPE_CUT_BASIC, new ModelResourceLocation(RS.ID + ":cut_basic_processor", "inventory"));
-        modelRegistration.setModel(this, TYPE_CUT_IMPROVED, new ModelResourceLocation(RS.ID + ":cut_improved_processor", "inventory"));
-        modelRegistration.setModel(this, TYPE_CUT_ADVANCED, new ModelResourceLocation(RS.ID + ":cut_advanced_processor", "inventory"));
+        modelRegistration.setModel(this, TYPE_CUT_BASIC, new ModelResourceLocation(RS.ID + ":raw_basic_processor", "inventory"));
+        modelRegistration.setModel(this, TYPE_CUT_IMPROVED, new ModelResourceLocation(RS.ID + ":raw_improved_processor", "inventory"));
+        modelRegistration.setModel(this, TYPE_CUT_ADVANCED, new ModelResourceLocation(RS.ID + ":raw_advanced_processor", "inventory"));
         modelRegistration.setModel(this, TYPE_BASIC, new ModelResourceLocation(RS.ID + ":basic_processor", "inventory"));
         modelRegistration.setModel(this, TYPE_IMPROVED, new ModelResourceLocation(RS.ID + ":improved_processor", "inventory"));
         modelRegistration.setModel(this, TYPE_ADVANCED, new ModelResourceLocation(RS.ID + ":advanced_processor", "inventory"));
-        modelRegistration.setModel(this, TYPE_CUT_SILICON, new ModelResourceLocation(RS.ID + ":cut_silicon", "inventory"));
     }
 
     @Override
@@ -56,7 +53,7 @@ public class ItemProcessor extends ItemBase {
             return;
         }
 
-        for (int i = 0; i <= 6; ++i) {
+        for (int i = 0; i <= 5; ++i) {
             items.add(new ItemStack(this, 1, i));
         }
     }