+++ src/main/java/com/raoulvdberge/refinedstorage/item/ItemProcessorBinding.java
@@ -1,0 +1,10 @@
+package com.raoulvdberge.refinedstorage.item;
+
+import com.raoulvdberge.refinedstorage.RS;
+import com.raoulvdberge.refinedstorage.item.info.ItemInfo;
+
+public class ItemProcessorBinding extends ItemBase {
+    public ItemProcessorBinding() {
+        super(new ItemInfo(RS.ID, "processor_binding"));
+    }
+}